Refactor keymap into own component
