+++ packages/web-app-files/src/App.vue
@@ -1,5 +1,6 @@
 <template>
   <main id="files" class="oc-flex oc-height-1-1">
+    <keymap-actions />
     <div ref="filesListWrapper" tabindex="-1" class="files-list-wrapper oc-width-expand">
       <router-view id="files-view" />
     </div>
@@ -19,29 +20,24 @@
   </main>
 </template>
 <script lang="ts">
-import { bus } from 'web-pkg/src/instance'
 import Mixins from './mixins'
 import { mapActions, mapState, mapMutations } from 'vuex'
 import SideBar from './components/SideBar/SideBar.vue'
+import KeymapActions from './components/FilesList/KeymapActions.vue'
 import { defineComponent } from '@vue/composition-api'
 import { usePublicLinkPassword, useStore } from 'web-pkg/src/composables'
 
 export default defineComponent({
   components: {
-    SideBar
+    SideBar,
+    KeymapActions
   },
   mixins: [Mixins],
-  data: () => {
-    return {
-      selectionCounter: 0
-    }
-  },
   computed: {
     ...mapState('Files/sidebar', {
       sidebarClosed: 'closed',
       sidebarActivePanel: 'activePanel'
     }),
-    ...mapState('Files', ['latestSelectedId']),
 
     showSidebar() {
       return !this.sidebarClosed
@@ -62,104 +58,14 @@ export default defineComponent({
     this.$root.$on('upload-end', () => {
       this.delayForScreenreader(() => this.$refs.filesListWrapper.focus())
     })
-    const fileListClickedEvent = bus.subscribe('app.files.list.clicked', (resource) => {
-      console.log(resource);
-    })
-
-    this.$on('beforeDestroy', () => {
-      bus.unsubscribe('app.files.list.clicked', fileListClickedEvent)
-    })
-  },
-
-  mounted() {
-    document.addEventListener('keydown', this.handleShortcut, false)
-  },
-
-  beforeDestroy() {
-    document.removeEventListener('keydown', this.handleShortcut)
   },
 
   methods: {
-    ...mapActions('Files', ['resetFileSelection', 'toggleFileSelection']),
+    ...mapActions('Files', ['resetFileSelection']),
     ...mapActions('Files/sidebar', {
       closeSidebar: 'close',
       setActiveSidebarPanel: 'setActivePanel'
     }),
-    ...mapActions(['showMessage', 'createModal', 'hideModal']),
-    ...mapActions('Files', ['copySelectedFiles', 'cutSelectedFiles', 'pasteSelectedFiles']),
-    ...mapMutations('Files', ['UPSERT_RESOURCE', 'SET_LATEST_SELECTED_FILE']),
-
-    handleShortcut(event) {
-      this.handleFileActionsShortcuts(event)
-      this.handleFileSelectionShortcuts(event)
-    },
-
-    handleFileSelectionShortcuts(event) {
-      // click has to reset selectioncounter
-      const key = event.keyCode || event.which
-      const isShiftPressed = event.shiftKey
-      if(!isShiftPressed) return
-      const isUpPressed = key === 38
-      const isDownPressed = key === 40
-      if(isDownPressed) {
-        const latestSelectedRow = document.querySelectorAll(`[data-item-id='${this.latestSelectedId}']`)[0]
-        const nextRow = latestSelectedRow.nextSibling as HTMLElement
-        if(nextRow === null) return
-        const nextResourceId = nextRow.getAttribute("data-item-id")
-
-        if(this.selectionCounter < 0) {
-          // deselect
-          this.toggleFileSelection({id: this.latestSelectedId})
-          this.SET_LATEST_SELECTED_FILE(nextResourceId)
-        }else {
-          // select
-          this.toggleFileSelection({id: nextResourceId})
-        }
-        this.selectionCounter = this.selectionCounter + 1
-      }
-      if(isUpPressed) {
-        const latestSelectedRow = document.querySelectorAll(`[data-item-id='${this.latestSelectedId}']`)[0]
-        const nextRow = latestSelectedRow.previousSibling as HTMLElement
-        if(nextRow === null) return
-        const nextResourceId = nextRow.getAttribute("data-item-id")
-        if(this.selectionCounter > 0) {
-          // deselect
-          this.toggleFileSelection({id: this.latestSelectedId})
-          this.SET_LATEST_SELECTED_FILE(nextResourceId)
-        }else {
-          // select
-          this.toggleFileSelection({id: nextResourceId})
-        }
-        this.selectionCounter = this.selectionCounter - 1
-      }
-    },
-
-    handleFileActionsShortcuts(event) {
-      const key = event.keyCode || event.which
-      const ctr = window.navigator.platform.match('Mac') ? event.metaKey : event.ctrlKey
-      if (!ctr /* CTRL | CMD */) return
-      const isCopyAction = key === 67
-      const isPasteAction = key === 86
-      const isCutAction = key === 88
-      if (isCopyAction) {
-        this.copySelectedFiles()
-      } else if (isPasteAction) {
-        this.pasteSelectedFiles({
-          client: this.$client,
-          createModal: this.createModal,
-          hideModal: this.hideModal,
-          showMessage: this.showMessage,
-          $gettext: this.$gettext,
-          $gettextInterpolate: this.$gettextInterpolate,
-          $ngettext: this.$ngettext,
-          routeContext: this.$route.name,
-          publicLinkPassword: this.publicLinkPassword,
-          upsertResource: this.UPSERT_RESOURCE
-        })
-      } else if (isCutAction) {
-        this.cutSelectedFiles()
-      }
-    },
 
     focusSideBar(component, event) {
       this.focus({
